Cherry pick commits in #6601 to main (#6611)
* fix write


* decoding ctc


* temp set rnnt decoding return_best_hypothesis to true


* add wer cal back to transcribe_speech as requested


* add wer cal back to speech_to_text_buffered_infer_rnnt  as requested


* add wer cal back to speech_to_text_buffered_infer_ctc as requested


* style fix


* reflect change in asr_evaluator


* reflect som and vahid comment


* remove return_best_hy=true in transcribe_speech


* no text skip


* revert partial


---------
